+++ CleverPush/Source/CPBannerCardContainer.m
@@ -537,6 +537,9 @@
     dispatch_async(dispatch_get_main_queue(), ^(void) {
         [[NSUserDefaults standardUserDefaults] setBool:false forKey:CLEVERPUSH_APP_BANNER_VISIBLE_KEY];
         [[NSUserDefaults standardUserDefaults] synchronize];
+        if (self.handleBannerClosed) {
+            self.handleBannerClosed();
+        }
         [self.controller dismissViewControllerAnimated:NO completion:nil];
     });
 }