+++ CleverPush/Source/CleverPush.m
@@ -576,6 +576,10 @@ static CleverPush* singleInstance = nil;
     [self.CPSharedInstance showAppBanner:bannerId];
 }
 
++ (void)showAppBanner:(NSString* _Nullable)bannerId appBannerClosedCallback:(CPAppBannerClosedBlock _Nullable)appBannerClosedCallback {
+    [self.CPSharedInstance showAppBanner:bannerId appBannerClosedCallback:appBannerClosedCallback];
+}
+
 + (void)setAppBannerOpenedCallback:(CPAppBannerActionBlock _Nullable)callback {
     [self.CPSharedInstance setAppBannerOpenedCallback:^(CPAppBannerAction*action) {
         callback(action);